// src/customer-preference/customer-preference.enums.ts
export enum CustomerLanguage {
  EN = 'EN',
  FR = 'FR',
  HA = 'HA',
  IG = 'IG',
  YO = 'YO',
}

export enum CustomerTheme {
  SYSTEM = 'SYSTEM',
  LIGHT = 'LIGHT',
  DARK = 'DARK',
}

export enum NotificationChannel {
  EMAIL = 'EMAIL',
  SMS = 'SMS',
  PUSH = 'PUSH',
  IN_APP = 'IN_APP',
}

export enum SecuritySetting {
  LOGIN_ALERTS = 'LOGIN_ALERTS',
  TRANSACTION_ALERTS = 'TRANSACTION_ALERTS',
  DEVICE_REGISTRATION_ALERTS = 'DEVICE_REGISTRATION_ALERTS',
  BIOMETRIC_ALLOWED = 'BIOMETRIC_ALLOWED',
}

export enum PreferenceHistoryAction {
  CREATED = 'CREATED',
  UPDATED = 'UPDATED',
}